# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Accessibility API",
    "version": "0.0.0"
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session"
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "error",
          "message"
        ]
      },
...
      "Statement": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "siteId": {
            "type": "string",
            "format": "uuid"
          },
          "version": {
            "type": "integer",
            "minimum": 1
          },
          "status": {
            "$ref": "#/components/schemas/StatementStatus"
          },
          "computedStatus": {
            "$ref": "#/components/schemas/ComputedComplianceStatus"
          },
          "declaredStatus": {
            "type": [
              "string",
              "null"
            ],
            "enum": [
              "total",
              "partiel",
              "non",
              null
            ]
          },
          "complianceRate": {
            "type": [
              "integer",
              "null"
            ],
            "minimum": 0,
            "maximum": 100
          },
          "allowedStatuses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ComplianceStatus"
            }
          },
          "counts": {
            "type": "object",
            "properties": {
              "conforme": {
                "type": "integer",
                "minimum": 0
              },
              "nonConforme": {
                "type": "integer",
                "minimum": 0
              },
              "na": {
                "type": "integer",
                "minimum": 0
              },
              "aVerifier": {
                "type": "integer",
                "minimum": 0
              }
            },
            "required": [
              "conforme",
              "nonConforme",
              "na",
              "aVerifier"
            ]
          },
          "unattributed": {
            "type": "integer",
            "minimum": 0
          },
          "entityName": {
            "type": [
              "string",
              "null"
            ]
          },
          "contactEmail": {
            "type": [
              "string",
              "null"
            ]
          },
          "contactUrl": {
            "type": [
              "string",
              "null"
            ]
          },
          "derogations": {
            "type": [
              "string",
              "null"
            ]
          },
          "samplePages": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "technologies": {
            "type": [
              "string",
              "null"
            ]
          },
          "testEnvironment": {
            "type": [
              "string",
              "null"
            ]
          },
          "tools": {
            "type": [
              "string",
              "null"
            ]
          },
          "nonAccessibleContent": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "criterionId": {
                  "type": "string"
                },
                "title": {
                  "type": "string"
                },
                "sources": {
                  "type": "array",
                  "items": {
                    "type": "string",
                    "enum": [
                      "manual",
                      "auto"
                    ]
                  }
                },
                "notes": {
                  "type": "string"
                }
              },
              "required": [
                "criterionId",
                "title",
                "sources",
                "notes"
              ]
            }
          },
          "locale": {
            "type": "string"
          },
          "referentialVersion": {
            "type": "string"
          },
          "missing": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "publicPath": {
            "type": [
              "string",
              "null"
            ]
          },
          "pdfReady": {
            "type": "boolean"
          },
          "publishedAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "siteId",
          "version",
          "status",
          "computedStatus",
          "declaredStatus",
          "complianceRate",
          "allowedStatuses",
          "counts",
          "unattributed",
          "entityName",
          "contactEmail",
          "contactUrl",
          "derogations",
          "samplePages",
          "technologies",
          "testEnvironment",
          "tools",
          "nonAccessibleContent",
          "locale",
          "referentialVersion",
          "missing",
          "publicPath",
          "pdfReady",
          "publishedAt",
          "createdAt",
          "updatedAt"
        ]
      },
      "StatementStatus": {
        "type": "string",
        "enum": [
          "draft",
          "published",
          "superseded"
        ]
      },
      "ComputedComplianceStatus": {
        "type": "string",
        "enum": [
          "total",
          "partiel",
          "non",
          "indetermine"
        ]
      },
      "ComplianceStatus": {
        "type": "string",
        "enum": [
          "total",
          "partiel",
          "non"
        ]
      },
      "UpdateStatementRequest": {
        "type": "object",
        "properties": {
          "entityName": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 200
          },
          "contactEmail": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 200,
            "format": "email"
          },
          "contactUrl": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 2000,
            "format": "uri"
          },
          "derogations": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 2000
          },
          "samplePages": {
            "type": "array",
            "items": {
              "type": "string",
              "maxLength": 2000,
              "format": "uri"
            },
            "maxItems": 50
          },
          "technologies": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 2000
          },
          "testEnvironment": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 2000
          },
          "tools": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 2000
          }
        }
      },
      "PublishStatementRequest": {
        "type": "object",
        "properties": {
          "declaredStatus": {
            "$ref": "#/components/schemas/ComplianceStatus"
          }
        },
        "required": [
          "declaredStatus"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/sites/{id}/statements": {
      "post": {
        "summary": "Start a draft accessibility statement from the current audit",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "201": {
            "description": "Draft created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Statement"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "A draft already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Versions of the accessibility statement, newest first",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Statements",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Statement"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not signed in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/statements/{id}": {
      "get": {
        "summary": "One statement (a draft is computed from the live audit)",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Statement",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Statement"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Edit a draft statement",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStatementRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Statement"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Not a draft",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/statements/{id}/publish": {
      "post": {
        "summary": "Publish a draft: the page becomes public, the previous version is superseded",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublishStatementRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Published",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Statement"
                }
              }
            }
          },
          "400": {
            "description": "Incomplete or level not supported by the audit",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Only an owner may publish",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Not a draft",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}